requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=4.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
streamlit>=1.40.0
